+++ src/components/About/AboutView.tsx
@@ -8,6 +8,7 @@ import { DataFoundationSection } from './sections/DataFoundationSection'
 import { SbomSection } from './sections/SbomSection'
 import { SecurityAuditSection } from './sections/SecurityAuditSection'
 import { DataPrivacySection } from './sections/DataPrivacySection'
+import { EmbeddingModeSection } from './sections/EmbeddingModeSection'
 import { LicenseSection } from './sections/LicenseSection'
 import { RagAiSection } from './sections/RagAiSection'
 import { CryptoBuffSection } from './sections/CryptoBuffSection'
@@ -27,6 +28,7 @@ export function AboutView() {
       <SbomSection />
       <SecurityAuditSection />
       <DataPrivacySection />
+      <EmbeddingModeSection />
       <LicenseSection />
       <RagAiSection />
       <CryptoBuffSection />